Consolidate PATH lookup: replace duplicate impls with shebang.LookPathInEnv
exec.LookPath is not suitable here because it uses the current process PATH
rather than an explicit PATH string; the verification path must resolve
commands in the runtime environment's PATH, which may differ.

- Export shebang.envLookPath as shebang.LookPathInEnv (with doc explaining
  why exec.LookPath cannot be used as a drop-in replacement)
- Rename private helpers envContainsPathSeparator/envIsExecutableFile to
  containsPathSeparator/isExecutableFile
- Delete identical lookPathInEnv/containsPathSeparator/isExecutableFile
  from verification/manager.go
- Update verification/path_resolver.go to call shebang.LookPathInEnv,
  translating shebang.ErrCommandNotFound → verification.ErrCommandNotFound
