+++ internal/verification/manager.go
@@ -780,47 +780,3 @@ func (m *Manager) verifyEnvPathResolution(commandName, recordedResolvedPath stri
 	}
 	return nil
 }
-
-// lookPathInEnv searches for an executable named name in the directories listed
-// in pathEnv (colon-separated on UNIX, semicolon-separated on Windows).
-// Returns the first matching executable or ErrCommandNotFound.
-func lookPathInEnv(name, pathEnv string) (string, error) {
-	if containsPathSeparator(name) {
-		// name itself is a relative or absolute path — use directly.
-		if isExecutableFile(name) {
-			return name, nil
-		}
-		return "", ErrCommandNotFound
-	}
-
-	for _, dir := range filepath.SplitList(pathEnv) {
-		if dir == "" {
-			dir = "."
-		}
-		candidate := filepath.Join(dir, name)
-		if isExecutableFile(candidate) {
-			return candidate, nil
-		}
-	}
-	return "", ErrCommandNotFound
-}
-
-// containsPathSeparator reports whether name contains a filepath separator.
-func containsPathSeparator(name string) bool {
-	for _, c := range name {
-		if c == '/' || c == filepath.Separator {
-			return true
-		}
-	}
-	return false
-}
-
-// isExecutableFile reports whether path names a regular file with at least one
-// execute bit set.
-func isExecutableFile(path string) bool {
-	info, err := os.Stat(path)
-	if err != nil {
-		return false
-	}
-	return info.Mode().IsRegular() && info.Mode().Perm()&0o111 != 0
-}
